requirements.txt
dj-database-url==0.4.1
Django==1.9.5
django-bootstrap3==7.0.1
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-rest-swagger==0.3.5
djangorestframework==3.3.3
gunicorn==19.4.5
httplib2==0.9.2
oauth2==1.9.0.post1
Pillow==3.2.0
postgres==2.2.1
psycopg2==2.6.1
python-dateutil==2.5.2
PyYAML==3.11
requests==2.9.1
six==1.10.0
sqlparse==0.1.19
stripe==1.32.2
Werkzeug==0.11.5
whitenoise==3.0
